# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PureX Video Aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface Video {
  id: number;
  slug: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  embed_url?: string | null;
  /** @nullable */
  external_id?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  duration_seconds?: number | null;
  /** @nullable */
  category_id?: number | null;
  /** @nullable */
  category_name?: string | null;
  status: string;
  views_count: number;
  likes_count: number;
  trending_score: number;
  is_editor_pick: boolean;
  tags: string[];
  performers?: string[];
  /** @nullable */
  studio?: string | null;
  /** @nullable */
  published_at?: string | null;
  created_at: string;
  /** @nullable */
  submitter_id?: number | null;
}

export interface VideoInput {
  title: string;
  description?: string;
  embed_code: string;
  thumbnail_url?: string;
}

export interface VideoStatusUpdate {
  status: string;
  is_editor_pick?: boolean;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  color?: string | null;
  sort_order: number;
  video_count?: number;
}

export interface SearchSuggestion {
  type: string;
  value: string;
}

export interface PerformerPage {
  name: string;
  total_views: number;
  video_count: number;
  videos: Video[];
}

export interface AuthInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  role: string;
  created_at: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface AdminStats {
  total_videos: number;
  published_videos: number;
  pending_videos: number;
  total_views: number;
  total_users: number;
}

export type ListVideosParams = {
limit?: number;
offset?: number;
category?: string;
source?: string;
order?: string;
q?: string;
};

export type GetFeaturedVideosParams = {
limit?: number;
};

export type GetTopVideosParams = {
limit?: number;
};

export type SearchSuggestParams = {
q: string;
};

export type FetchPerformerParams = {
name: string;
limit?: number;
offset?: number;
};

export type FetchStudioParams = {
name: string;
limit?: number;
offset?: number;
};

export type AdminListVideosParams = {
status?: string;
limit?: number;
offset?: number;
};