# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PureX Video Aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total_videos: number;
  published_videos: number;
  pending_videos: number;
  total_views: number;
  total_users: number;
}