# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PureX Video Aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthInput {
  email: string;
  password: string;
}