# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PureX Video Aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  slug: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  embed_url?: string | null;
  /** @nullable */
  external_id?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  duration_seconds?: number | null;
  /** @nullable */
  category_id?: number | null;
  /** @nullable */
  category_name?: string | null;
  status: string;
  views_count: number;
  likes_count: number;
  trending_score: number;
  is_editor_pick: boolean;
  tags: string[];
  performers?: string[];
  /** @nullable */
  studio?: string | null;
  /** @nullable */
  published_at?: string | null;
  created_at: string;
  /** @nullable */
  submitter_id?: number | null;
}